pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

packages = [
    { include = "myapp" },
]

[tool.poetry.scripts]
my-grpc-server = "myapp.server:main"
my-grpc-client = "myapp.client:main"

[tool.poetry.dependencies]
python = "^3.10"
grpcio-tools = "^1.43.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
